Add request summary to the request modal
